ai/requirements.txt
asyncpg==0.30.0
fastapi==0.116.1
matplotlib==3.10.5
nltk==3.9.1
numpy>=1.26.0,<2.1.0
pydantic==2.11.7
python-dotenv==1.1.1
seaborn==0.13.2
tensorflow==2.18.0
uvicorn==0.35.0
